+++ packages/colony-js-contract-loader-http/src/flowtypes.js
@@ -5,4 +5,5 @@ import type { Transform } from '@colony/colony-js-contract-loader';
 export type ConstructorArgs = {
   endpoint: string,
   transform: Transform,
+  network?: string,
 };